backend/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "", email = ""},
]
dependencies = [
    "Django==4.1",
    "djangorestframework==3.14",
    "django-environ>=0.11.2",
    "django-debug-toolbar>=4.2.0",
    "django-extensions>=3.2.3",
    "django-cors-headers>=4.3.1",
    "whitenoise>=6.6.0",
    "psycopg2-binary>=2.9.9",
    "drf-yasg>=1.21.7",
    "django-filter>=23.5",
    "djangorestframework-simplejwt>=5.3.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool.pdm.dev-dependencies]
dev = [
    "factory-boy>=3.3.0",
    "ipdb>=0.13.13",
    "pytest>=8.1.1",
    "pytest-django>=4.8.0",
    "pytest-dotenv>=0.5.2",
    "pytest-factoryboy>=2.7.0",
    "pytest-faker>=2.0.0",
    "pytest-freezegun>=0.4.2",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "testpath>=0.6.0",
]
